requirements.txt
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
tavily-python>=0.3.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
